Revert "Synchronize codes for OxygenOS 4.1.7"
This reverts commit 78f50f8bfae019b297d9c6df319c80dd226a2f35.

+++ sound/soc/msm/qdsp6v2/q6voice.h
@@ -141,7 +141,7 @@ struct share_mem_buf {
 struct mem_map_table {
 	dma_addr_t		phys;
 	void			*data;
-	size_t			size; /* size of buffer */
+	uint32_t		size; /* size of buffer */
 	struct ion_handle	*handle;
 	struct ion_client	*client;
 };
